qc/pipeline: relocate tests for 'core' functionality.
