--- auto_process_ngs/test/qc/pipeline/test_verify_fastqs.py
@@ -1,146 +1,0 @@
-#######################################################################
-# Unit tests for qc/pipeline.py (Fastq verification)
-#######################################################################
-
-# All imports declared in __init__.py file
-from . import *
-
-class TestQCPipeline(BaseQCPipelineTestCase):
-    """
-    Tests for Fastq verification
-    """
-    def test_qcpipeline_verify_fastqs_with_valid_fastqs(self):
-        """
-        QCPipeline: verify Fastqs with valid files
-        """
-        # Make mock QC executables
-        MockFastqScreen.create(os.path.join(self.bin,"fastq_screen"))
-        MockFastQC.create(os.path.join(self.bin,"fastqc"))
-        MockStar.create(os.path.join(self.bin,"STAR"))
-        MockSamtools.create(os.path.join(self.bin,"samtools"))
-        MockPicard.create(os.path.join(self.bin,"picard"))
-        MockGtf2bed.create(os.path.join(self.bin,"gtf2bed"))
-        MockRSeQC.create(os.path.join(self.bin,"infer_experiment.py"))
-        MockRSeQC.create(os.path.join(self.bin,"geneBody_coverage.py"))
-        MockQualimap.create(os.path.join(self.bin,"qualimap"))
-        MockMultiQC.create(os.path.join(self.bin,"multiqc"))
-        os.environ['PATH'] = "%s:%s" % (self.bin,
-                                        os.environ['PATH'])
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={ 'Organism': 'Human' })
-        p.create(top_dir=self.wd)
-        # Set up and run the QC
-        runqc = QCPipeline()
-        runqc.add_project(AnalysisProject(os.path.join(self.wd,"PJB")),
-                          fetch_protocol_definition("standardPE"),
-                          verify_fastqs=True,
-                          multiqc=True)
-        status = runqc.run(fastq_screens=self.fastq_screens,
-                           star_indexes=
-                           { 'human': '/data/hg38/star_index' },
-                           annotation_bed_files=
-                           { 'human': self.ref_data['hg38']['bed'] },
-                           annotation_gtf_files=
-                           { 'human': self.ref_data['hg38']['gtf'] },
-                           poll_interval=POLL_INTERVAL,
-                           max_jobs=1,
-                           runners={ 'default': SimpleJobRunner(), })
-        self.assertEqual(status,0)
-        # Check QC metadata
-        qc_info = AnalysisProjectQCDirInfo(
-            os.path.join(self.wd,"PJB","qc","qc.info"))
-        self.assertEqual(qc_info.protocol,"standardPE")
-        self.assertEqual(qc_info.protocol_specification,
-                         str(fetch_protocol_definition("standardPE")))
-        self.assertEqual(qc_info.organism,"Human")
-        self.assertEqual(qc_info.fastq_dir,
-                         os.path.join(self.wd,"PJB","fastqs"))
-        self.assertEqual(qc_info.fastqs,
-                         "PJB1_S1_R1_001.fastq.gz,"
-                         "PJB1_S1_R2_001.fastq.gz,"
-                         "PJB2_S2_R1_001.fastq.gz,"
-                         "PJB2_S2_R2_001.fastq.gz")
-        self.assertEqual(qc_info.fastqs_split_by_lane,False)
-        self.assertEqual(qc_info.fastq_screens,
-                         "model_organisms,other_organisms,rRNA")
-        self.assertEqual(qc_info.star_index,"/data/hg38/star_index")
-        self.assertEqual(qc_info.annotation_bed,self.ref_data['hg38']['bed'])
-        self.assertEqual(qc_info.annotation_gtf,self.ref_data['hg38']['gtf'])
-        self.assertEqual(qc_info.cellranger_version,None)
-        self.assertEqual(qc_info.cellranger_refdata,None)
-        self.assertEqual(qc_info.cellranger_probeset,None)
-
-    def test_qcpipeline_verify_fastqs_with_corrupted_fastq(self):
-        """
-        QCPipeline: verify Fastqs with corrupted file
-        """
-        # Make mock QC executables
-        MockFastqScreen.create(os.path.join(self.bin,"fastq_screen"))
-        MockFastQC.create(os.path.join(self.bin,"fastqc"))
-        MockStar.create(os.path.join(self.bin,"STAR"))
-        MockSamtools.create(os.path.join(self.bin,"samtools"))
-        MockPicard.create(os.path.join(self.bin,"picard"))
-        MockGtf2bed.create(os.path.join(self.bin,"gtf2bed"))
-        MockRSeQC.create(os.path.join(self.bin,"infer_experiment.py"))
-        MockRSeQC.create(os.path.join(self.bin,"geneBody_coverage.py"))
-        MockQualimap.create(os.path.join(self.bin,"qualimap"))
-        MockMultiQC.create(os.path.join(self.bin,"multiqc"))
-        os.environ['PATH'] = "%s:%s" % (self.bin,
-                                        os.environ['PATH'])
-        # Make mock analysis project
-        p = MockAnalysisProject("PJB",("PJB1_S1_R1_001.fastq.gz",
-                                       "PJB1_S1_R2_001.fastq.gz",
-                                       "PJB2_S2_R1_001.fastq.gz",
-                                       "PJB2_S2_R2_001.fastq.gz"),
-                                metadata={ 'Organism': 'Human' })
-        p.create(top_dir=self.wd)
-        # Replace a Fastq file with 'bad' version
-        bad_fastq = os.path.join(self.wd,
-                                 "PJB","fastqs",
-                                 "PJB1_S1_R2_001.fastq.gz")
-        with open(bad_fastq,'wb') as fp:
-            fp.write(b"Corrupted\ndata\nfor\nGzipped file!")
-        # Set up and run the QC
-        runqc = QCPipeline()
-        runqc.add_project(AnalysisProject(os.path.join(self.wd,"PJB")),
-                          fetch_protocol_definition("standardPE"),
-                          verify_fastqs=True,
-                          multiqc=True)
-        status = runqc.run(fastq_screens=self.fastq_screens,
-                           star_indexes=
-                           { 'human': '/data/hg38/star_index' },
-                           annotation_bed_files=
-                           { 'human': self.ref_data['hg38']['bed'] },
-                           annotation_gtf_files=
-                           { 'human': self.ref_data['hg38']['gtf'] },
-                           poll_interval=POLL_INTERVAL,
-                           max_jobs=1,
-                           runners={ 'default': SimpleJobRunner(), })
-        self.assertEqual(status,1)
-        # Check QC metadata
-        qc_info = AnalysisProjectQCDirInfo(
-            os.path.join(self.wd,"PJB","qc","qc.info"))
-        self.assertEqual(qc_info.protocol,"standardPE")
-        self.assertEqual(qc_info.protocol_specification,
-                         str(fetch_protocol_definition("standardPE")))
-        self.assertEqual(qc_info.organism,"Human")
-        self.assertEqual(qc_info.fastq_dir,
-                         os.path.join(self.wd,"PJB","fastqs"))
-        self.assertEqual(qc_info.fastqs,
-                         "PJB1_S1_R1_001.fastq.gz,"
-                         "PJB1_S1_R2_001.fastq.gz,"
-                         "PJB2_S2_R1_001.fastq.gz,"
-                         "PJB2_S2_R2_001.fastq.gz")
-        self.assertEqual(qc_info.fastqs_split_by_lane,False)
-        self.assertEqual(qc_info.fastq_screens,
-                         "model_organisms,other_organisms,rRNA")
-        self.assertEqual(qc_info.star_index,"/data/hg38/star_index")
-        self.assertEqual(qc_info.annotation_bed,self.ref_data['hg38']['bed'])
-        self.assertEqual(qc_info.annotation_gtf,self.ref_data['hg38']['gtf'])
-        self.assertEqual(qc_info.cellranger_version,None)
-        self.assertEqual(qc_info.cellranger_refdata,None)
-        self.assertEqual(qc_info.cellranger_probeset,None)
